feat: implement bulk update handlers for anime, manga, shows, and podcasts
- Added utility functions for handling bulk updates for different media types in `bulk-update-handlers.ts`.
- Created common UI components for metadata progress updates in `common-elements.tsx`.
- Defined types for metadata progress forms in `form-types.ts`.
- Refactored the dashboard to use the new forms structure and removed the old metadata progress update forms.
