+++ apps/frontend/app/components/dashboard/forms/metadata-progress-update/in-progress-form.tsx
@@ -1,0 +1,130 @@
+import {
+	Button,
+	Flex,
+	Group,
+	NumberInput,
+	Slider,
+	Stack,
+	Text,
+} from "@mantine/core";
+import { MediaLot } from "@ryot/generated/graphql/backend/graphql";
+import { isNumber } from "@ryot/ts-utils";
+import {
+	IconBook,
+	IconBrandPagekit,
+	IconClock,
+	IconDeviceTv,
+	IconPercentage,
+} from "@tabler/icons-react";
+import { useState } from "react";
+import { match } from "ts-pattern";
+import { useDeployBulkMetadataProgressUpdate } from "~/lib/hooks";
+import type { MetadataInProgressFormProps } from "./utils/form-types";
+
+export const MetadataInProgressUpdateForm = ({
+	onSubmit,
+	inProgress,
+	metadataDetails,
+	metadataToUpdate,
+}: MetadataInProgressFormProps) => {
+	const deployBulkMetadataProgressUpdate = useDeployBulkMetadataProgressUpdate(
+		metadataDetails.title,
+	);
+
+	const total =
+		metadataDetails.audioBookSpecifics?.runtime ||
+		metadataDetails.bookSpecifics?.pages ||
+		metadataDetails.movieSpecifics?.runtime ||
+		metadataDetails.mangaSpecifics?.chapters ||
+		metadataDetails.animeSpecifics?.episodes ||
+		metadataDetails.visualNovelSpecifics?.length;
+	const progress = Number(inProgress.progress);
+	const [value, setValue] = useState<number | undefined>(progress);
+
+	const [updateIcon, text] = match(metadataDetails.lot)
+		.with(MediaLot.Book, () => [<IconBook size={24} key="element" />, "Pages"])
+		.with(MediaLot.Anime, () => [
+			<IconDeviceTv size={24} key="element" />,
+			"Episodes",
+		])
+		.with(MediaLot.Manga, () => [
+			<IconBrandPagekit size={24} key="element" />,
+			"Chapters",
+		])
+		.with(MediaLot.Movie, MediaLot.VisualNovel, MediaLot.AudioBook, () => [
+			<IconClock size={24} key="element" />,
+			"Minutes",
+		])
+		.otherwise(() => [null, null]);
+
+	return (
+		<Stack mt="sm">
+			<Group>
+				<Slider
+					min={0}
+					step={1}
+					max={100}
+					value={value}
+					onChange={setValue}
+					style={{ flexGrow: 1 }}
+					showLabelOnHover={false}
+				/>
+				<NumberInput
+					w="20%"
+					min={0}
+					step={1}
+					max={100}
+					hideControls
+					value={value}
+					onFocus={(e) => e.target.select()}
+					rightSection={<IconPercentage size={16} />}
+					onChange={(v) => {
+						if (isNumber(v)) setValue(v);
+						else setValue(undefined);
+					}}
+				/>
+			</Group>
+			{total ? (
+				<>
+					<Text ta="center" fw="bold">
+						OR
+					</Text>
+					<Flex align="center" gap="xs">
+						<NumberInput
+							min={0}
+							step={1}
+							flex={1}
+							hideControls
+							leftSection={updateIcon}
+							max={Number(total)}
+							onFocus={(e) => e.target.select()}
+							defaultValue={((Number(total) || 1) * (value || 1)) / 100}
+							onChange={(v) => {
+								const value = (Number(v) / (Number(total) || 1)) * 100;
+								setValue(value);
+							}}
+						/>
+						<Text>{text}</Text>
+					</Flex>
+				</>
+			) : null}
+			<Button
+				variant="outline"
+				type="submit"
+				onClick={async () => {
+					await deployBulkMetadataProgressUpdate.mutateAsync([
+						{
+							metadataId: metadataToUpdate.metadataId,
+							change: {
+								changeLatestInProgress: { progress: value?.toString() },
+							},
+						},
+					]);
+					onSubmit();
+				}}
+			>
+				Update
+			</Button>
+		</Stack>
+	);
+};
